requirements.txt
numpy==1.26.4
scikit-learn==1.2.2
imbalanced-learn==0.12.4
pandas==2.2.3
seaborn==0.13.2
matplotlib==3.10.0
mlflow==3.1.1
dvc==3.60.1
Flask==3.1.0
PyYAML
gunicorn
